requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
psycopg[binary]
python-dotenv
flasgger
flask-jwt-extended
werkzeug
gunicorn
cloudinary
# Paystack SDK
paystackapi==2.0.0